observables: allow subgroups, add "pressure", matrix of edge vectors
We shall allow subgroups in the "observables" group as we do in "trajectory".

We shall not forget the pressue, which is of similar importance as the
potential energy or the temperature (aka kinetic energy).

For the triclinic box, the set of D D-dimensional vectors may be considered a
D×D matrix which makes essentially no difference how the data are stored
finally. But it is clearer that the full set of edge vectors is to be
considered a single entity, e.g., the determinant of which readily yields the
box volume.