explain the interpretation of "particles/id"
The specs should specify exactly how the groups are to be interpreted,
and how it is to be interpreted if a group is missing. A general
disclaimer that any group that is missing is simply unspecified would
solve most cases. In the case of the "id", however, it is not.