2nd revision of thermodynamics module
* rename particles to particle_number

This avoids a potential confusion between the root group "particles" and
the H5MD element "particles". Further, the field stores no detailed
information about the particles (which somebody could expect from the
name), just the bare particle number.

* store intensive quantities (per particle) instead of extensive ones

This reverts a previous preference for extensive quantities. The latter,
however, can become really huge as they depend trivially on the system
size and are not very practical, for example, it is difficult to compare
them by visual inspection. Intensive quantities on the other hand do not
suffer from such issues and are typically quoted in publications. The
conversion to extensive quantities is always possible since the particle
number must be given as well.