units modules: drop UTF8 encoding
Restricting to the ASCII character set ensures best compatibility and
avoids possible complications when handling UTF8. The units field is
regarded as a data field and is primarily not for display purposes.

See the discussion starting at gmane.science.simulation.h5md.user/490

For each symbol, there is only a single variant now (no choices anymore
between µ and u or Å and Ang), °C has become degC in accordance with
udunits2.

Drop ohm, faraday, tesla. There is no satisfactory replacement for Ω,
siemens, weber, and henry were missing, and these units are easily build
from V, A, s, m.

Add that the integer indicating a power must be signed, which simplifies
parsing and avoids issues as in "103 m" actually meaning "10+3 m".

I kept the UTF8 in the description fields in the tables for better
readability.