particle/position: specify optional bounding box
In some situations there is a need to guarantee the bounds of the
particle positions, e.g., for visualisation. Add an optional attribute
to particle/…/position to specify such a bounding box.

The attribute supersedes and generalises the box origin (dropped in the
previous commit). It permits the situation that the particles are
outside the primary box by a (small) "skin width", although the data are
to be interpreted as periodically reduced positions.

For a discussion, see
http://article.gmane.org/gmane.science.simulation.h5md.user/516
http://article.gmane.org/gmane.science.simulation.h5md.user/523